refactor(session): extract SessionCapable trait; fix IBS\SessionClient LSP violation
IBS\SessionClient had four methods (login, logout, saveSession,
reuseSession) that unconditionally threw "Method not supported" — the
class was named SessionClient but incapable of session communication.

Fix: extract the real session implementations into a CNIC\CNR\SessionCapable
trait. CNR\SessionClient uses the trait; IBS\SessionClient does not,
so it no longer carries misleading stubs. MONIKER\SessionClient (which
extends IBS) is unaffected — it also does not support sessions.

The trait is the extension point for any future registrar that does
support sessions: include it and the capability is available.

Test adjusted: testSaveReuseSession instantiates SessionClient directly
rather than via ClientFactory so PHPStan can infer the precise type.
